requirements.txt
torch
torchvision
numpy
h5py
Pillow
tqdm
matplotlib
scikit-image
opencv-python
tensorboard
wandb
scipy
pandas
pyyaml
jsonschema